# docs/swagger.yaml
definitions:
...
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  request.PermissionRequest:
    properties:
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
      updated_by:
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Api Service boilerplate dari golang gin dan gorm
  termsOfService: http://swagger.io/terms/
  title: My Admin Service Api
  version: "1.0"
paths:
...
  /permissions:
    get:
      consumes:
      - application/json
      description: REST API for permission
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get list permission
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: REST API for Permissions
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.PermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Insert Permissions
      tags:
      - Permissions
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: REST API for Permissions
      parameters:
      - description: Permissions ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Delete Permissions
      tags:
      - Permissions
    get:
      consumes:
      - application/json
      description: REST API for Permissions
      parameters:
      - description: Permissions ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get one Permissions
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: REST API for Permissions
      parameters:
      - description: Permissions ID
        in: query
        name: id
        type: string
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.PermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Update Permissions
      tags:
      - Permissions
  /privileges:
    get:
      consumes:
      - application/json
      description: REST API for Privileges
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get list Privileges
      tags:
      - Privilege
  /privileges/{id}:
    get:
      consumes:
      - application/json
      description: REST API for Privileges
      parameters:
      - description: Privileges ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get one Privileges
      tags:
      - Privilege
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"